Extract podcast cover to compose components
(cherry picked from commit 4fa25832ba76a8a6d763015e77ecc2ad67d35fd8)
